refactor: extract TimerUI component from Timer
